Fix format price BigDecimal pada order dan product service

+++ product/src/main/java/com/example/product/model/Product.java
@@ -1,6 +1,7 @@
 package com.example.product.model;
 
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "products")
@@ -12,10 +13,9 @@ public class Product {
 
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price; 
     private Integer stock;
 
-    // Getter & Setter
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
 
@@ -25,8 +25,8 @@ public class Product {
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
 
-    public Double getPrice() { return price; }
-    public void setPrice(Double price) { this.price = price; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
 
     public Integer getStock() { return stock; }
     public void setStock(Integer stock) { this.stock = stock; }
